+++ admin/settings/settings-loader.php
@@ -28,6 +28,8 @@ function lmt_plug_settings_page() {
         if( count( $post_types ) >= 1 ) {
             add_settings_field('lmt_custom_post_types_list',  __( 'Enable Custom Post Types Support:', 'wp-last-modified-info' ), 'lmt_custom_post_types_list_display', 'lmt_post_option', 'lmt_post_option_section', array( 'label_for' => 'cpt', 'class' => 'cpt' ));  
         }
+        add_settings_field('lmt_gap_on_post',  __( 'Published Time & Modified Time Gap:', 'wp-last-modified-info' ), 'lmt_gap_on_post_display', 'lmt_post_option', 'lmt_post_option_section', array( 'label_for' => 'post-gap' ));  
+        
     // start page fields
     add_settings_section('lmt_page_option_section', __( 'Page Options', 'wp-last-modified-info' ) . '<p><hr></p>', null, 'lmt_page_option');    
         add_settings_field('lmt_enable_last_modified_page_cb', __( 'Enable for Pages on Frontend:', 'wp-last-modified-info' ), 'lmt_enable_last_modified_page_cb_display', 'lmt_page_option', 'lmt_page_option_section', array( 'label_for' => 'page-enable' ));  
@@ -37,6 +39,7 @@ function lmt_plug_settings_page() {
         add_settings_field('lmt_last_modified_format_page',  __( 'Last Modified Info Format for Pages:', 'wp-last-modified-info' ), 'lmt_last_modified_format_page_display', 'lmt_page_option', 'lmt_page_option_section', array( 'label_for' => 'page-format' ));  
         add_settings_field('lmt_last_modified_default_format_page',  __( 'Date and Time Visibility on Pages:', 'wp-last-modified-info' ), 'lmt_last_modified_default_format_page_display', 'lmt_page_option', 'lmt_page_option_section', array( 'label_for' => 'page-default-format', 'class' => 'page-default-format' ));  
         add_settings_field('lmt_show_author_page_cb', __( 'Display Last Modified Author Name:', 'wp-last-modified-info' ), 'lmt_show_author_page_cb_display', 'lmt_page_option', 'lmt_page_option_section', array( 'label_for' => 'page-sa' ));  
+        add_settings_field('lmt_gap_on_page',  __( 'Published Time & Modified Time Gap:', 'wp-last-modified-info' ), 'lmt_gap_on_page_display', 'lmt_page_option', 'lmt_page_option_section', array( 'label_for' => 'page-gap' ));  
         
     // start template tags
     add_settings_section('lmt_template_tag_section', __( 'Template Tags Options', 'wp-last-modified-info' ) . '<p><hr></p>', null, 'lmt_template_tag_option');